requirements.txt
gradio
tensorflow
numpy
pillow

